refactor(browser): colocate the public API plugin
