--- packages/browser/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@opencode-ai/browser",
-  "version": "0.0.0",
-  "type": "module",
-  "license": "MIT",
-  "exports": {
-    ".": "./src/index.ts",
-    "./rpc": "./src/rpc.ts"
-  },
-  "files": [
-    "src"
-  ],
-  "scripts": {
-    "test": "bun test --timeout 15000",
-    "typecheck": "tsgo --noEmit"
-  },
-  "peerDependencies": {
-    "@opencode-ai/plugin": "workspace:*",
-    "effect": "catalog:"
-  },
-  "devDependencies": {
-    "@opencode-ai/sdk": "workspace:*",
-    "@tsconfig/bun": "catalog:",
-    "@types/bun": "catalog:",
-    "@typescript/native-preview": "catalog:"
-  }
-}
