--- packages/browser/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@tsconfig/bun/tsconfig.json",
-  "compilerOptions": {
-    "lib": ["ESNext", "DOM", "DOM.Iterable"],
-    "strict": true,
-    "noUncheckedIndexedAccess": false,
-    "noEmit": true
-  },
-  "include": ["src", "test"]
-}
