+++ packages/schema/src/browser.ts
@@ -1,14 +1,15 @@
-export * as Browser from "./rpc.js"
+export * as Browser from "./browser.js"
 
-import { Session } from "@opencode-ai/plugin/effect"
-import { Rpc } from "@opencode-ai/plugin/rpc"
 import { Schema } from "effect"
+import { Rpc } from "./rpc.js"
+import { Session } from "./session.js"
 
 export const Ref = Schema.String.check(Schema.isPattern(/^@?e[1-9][0-9]*$/))
   .pipe(Schema.brand("Browser.Ref"))
   .annotate({ identifier: "Browser.Ref" })
 export type Ref = typeof Ref.Type
 
+export interface State extends Schema.Schema.Type<typeof State> {}
 export const State = Schema.Struct({
   url: Schema.String.check(Schema.isMaxLength(16_384)),
   title: Schema.String.check(Schema.isMaxLength(1_024)),
@@ -16,8 +17,7 @@ export const State = Schema.Struct({
   canGoBack: Schema.Boolean,
   canGoForward: Schema.Boolean,
   generation: Schema.Int.check(Schema.isGreaterThanOrEqualTo(0)),
-})
-export type State = typeof State.Type
+}).annotate({ identifier: "Browser.State" })
 
 export const Key = Schema.Literals([
   "Enter",
@@ -34,9 +34,9 @@ export const Key = Schema.Literals([
   "Home",
   "End",
   "Space",
-])
+]).annotate({ identifier: "Browser.Key" })
 export type Key = typeof Key.Type
-export const Direction = Schema.Literals(["up", "down", "left", "right"])
+export const Direction = Schema.Literals(["up", "down", "left", "right"]).annotate({ identifier: "Browser.Direction" })
 export type Direction = typeof Direction.Type
 
 export const Action = Schema.Union([
@@ -56,11 +56,13 @@ export const Action = Schema.Union([
     direction: Direction,
     pixels: Schema.Int.check(Schema.isGreaterThan(0), Schema.isLessThanOrEqualTo(2000)),
   }),
-])
+]).annotate({ identifier: "Browser.Action" })
 export type Action = typeof Action.Type
 
-export const Command = Schema.Struct({ action: Action, generation: State.fields.generation })
-export type Command = typeof Command.Type
+export interface Command extends Schema.Schema.Type<typeof Command> {}
+export const Command = Schema.Struct({ action: Action, generation: State.fields.generation }).annotate({
+  identifier: "Browser.Command",
+})
 export const Result = Schema.Union([
   Schema.Struct({ type: Schema.Literal("state"), state: State }),
   Schema.Struct({
@@ -78,12 +80,16 @@ export const Result = Schema.Union([
     state: State,
     data: Schema.Uint8ArrayFromBase64.check(Schema.isMaxLength(5 * 1_024 * 1_024)),
   }),
-]).pipe(Schema.toTaggedUnion("type"))
+])
+  .pipe(Schema.toTaggedUnion("type"))
+  .annotate({ identifier: "Browser.Result" })
 export type Result = typeof Result.Type
 export const Outcome = Schema.Union([
   Schema.Struct({ type: Schema.Literal("success"), result: Result }),
   Schema.Struct({ type: Schema.Literal("failure"), message: Schema.String.check(Schema.isMaxLength(1_024)) }),
-]).pipe(Schema.toTaggedUnion("type"))
+])
+  .pipe(Schema.toTaggedUnion("type"))
+  .annotate({ identifier: "Browser.Outcome" })
 export type Outcome = typeof Outcome.Type
 
 const attachment = { sessionID: Session.ID, connectionID: Schema.String }
@@ -97,7 +103,9 @@ export const Control = Schema.Union([
     command: Command,
   }),
   Schema.Struct({ type: Schema.Literal("cancel"), connectionID: Schema.String, requestID: Schema.String }),
-]).pipe(Schema.toTaggedUnion("type"))
+])
+  .pipe(Schema.toTaggedUnion("type"))
+  .annotate({ identifier: "Browser.Control" })
 export type Control = typeof Control.Type
 
 export const Definition = Rpc.define({
